\name{relateAdmix-package}
\alias{relateAdmix-package}
\alias{relateAdmix}
\docType{package}
\title{
Package for estimating relatedness for admixted individuals using genomic markers.
}
\description{
Package for estimating relatedness for admixted individuals using genomic markers.
}
\details{
\tabular{ll}{
Package: \tab relateAdmix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-12\cr
License: \tab GPL \cr
}
relate
}
\author{
Anders albrechtsen
Maintainer: Who to complain to <albrecht@binf.ku.dk>
}
\references{
Ida Moltke and Anders Albrechtsen
}
\keyword{ package }
\seealso{
Relate (www.popgen.dk/software)
}

